requirements.txt
agno
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
flask
groq
google-genai